Don't include vendor/autoload.php if S3Client is already loaded
When Mediawiki's composer.local.json file is used to install
dependencies for the AWS extension the AWS classes will already
be autoloaded when requested by the extension. Existing users
may still be running "composer install" within the extension
folder and generating an autoload.php file, so the absence of
the S3 classes will continue to triger this old behavior for
compatibility.